Reorganise js into dirs
