+++ src/rubintv/static/js/refresh-table/refresh-table-startracker.js
@@ -1,7 +1,7 @@
 /* global jQuery */
-import { parseJsonFromDOM, refreshTable } from './modules/table-control.js'
-import { addToTable } from './modules/table-startracker.js'
-import { starTrackerHeaders } from './models.js'
+import { parseJsonFromDOM, refreshTable } from '../modules/table-control.js'
+import { addToTable } from '../modules/table-startracker.js'
+import { starTrackerHeaders } from '../models.js'
 
 (function ($) {
   const meta = parseJsonFromDOM('#table-metadata')
