+++ src/rubintv/templates/cameras/allsky-historical.jinja
@@ -42,5 +42,5 @@
 {% endblock content %}
 {% block footer_scripts %}
 <script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
-<script type=module src="{{ url('static', filename='js/allsky-calendar.js') }}"></script>
+<script type=module src="{{ url('static', filename='js/allsky/allsky-calendar.js') }}"></script>
 {% endblock footer_scripts %}
