+++ src/rubintv/templates/cameras/allsky.jinja
@@ -32,5 +32,5 @@
 {% endblock content %}
 {% block footer_scripts %}
 <script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
-<script type=module src="{{ url('static', filename='js/allsky-refresh.js') }}"></script>
+<script type=module src="{{ url('static', filename='js/allsky/allsky-refresh.js') }}"></script>
 {% endblock footer_scripts %}
