+++ src/rubintv/templates/cameras/historical.jinja
@@ -12,7 +12,7 @@
 </section>
 {% endblock content %}
 {% block footer_scripts %}
-  {% set cam_js_filename = "js/camera-calendar-" + camera.slug + ".js" %}
+  {% set cam_js_filename = "js/camera-calendar/camera-calendar-" + camera.slug + ".js" %}
   <script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
   <script type=module src="{{ url('static', filename=cam_js_filename) }}"></script>
 {% endblock footer_scripts %}
